Churn Prediction
ID = Unique Identifier of a row

Age = Age of the customer

Gender = Gender of the customer (Male and Female)

Income = Yearly income of the customer

Balance = Average quarterly balance of the customer

Vintage = No. of years the customer is associated with bank

Transaction_Status = Whether the customer has done any transaction in the past 3 months or not

Product_Holdings = No. of product holdings with the bank

Credit_Card = Whether the customer has a credit card or not

Credit_Category = Category of a customer based on the credit score

Is_Churn = Whether the customer will churn in next 6 months or not